initial refactor of ops to kubefirst and clusters
