+++ clusters.go
@@ -1,0 +1,160 @@
+package main
+
+import (
+	"fmt"
+	"os"
+	"path/filepath"
+	"strings"
+
+	"github.com/charmbracelet/huh"
+	"github.com/charmbracelet/log"
+)
+
+func provisionCluster() {
+	log.Info("Starting provisionCluster function")
+	indexFile, err := loadIndexFile()
+	if err != nil {
+		log.Error("Error loading index file", "error", err)
+		fmt.Println("Failed to load configurations. Please ensure that the index.hcl file exists and is correctly formatted.")
+		return
+	}
+
+	log.Info("Index file loaded successfully", "version", indexFile.Version, "lastUpdated", indexFile.LastUpdated)
+	log.Info("Configs found", "count", len(indexFile.Configs))
+
+	// List available configs
+	var selectedConfig string
+	configOptions := make([]huh.Option[string], 0, len(indexFile.Configs))
+	for config, details := range indexFile.Configs {
+		log.Info("Config found", "name", config, "fileCount", len(details.Files))
+		configOptions = append(configOptions, huh.NewOption(config, config))
+	}
+
+	if len(configOptions) == 0 {
+		log.Warn("No configurations found in the index file")
+		fmt.Println("No configurations available. Please create a configuration first.")
+		fmt.Println("You can create a configuration using the 'Config' -> 'Create Config' option in the main menu.")
+		return
+	}
+
+	log.Info("Presenting config selection to user", "optionCount", len(configOptions))
+	form := huh.NewForm(
+		huh.NewGroup(
+			huh.NewSelect[string]().
+				Title("Select a configuration").
+				Options(configOptions...).
+				Value(&selectedConfig),
+		),
+	)
+
+	err = form.Run()
+	if err != nil {
+		log.Error("Error in config selection", "error", err)
+		return
+	}
+	log.Info("User selected config", "selectedConfig", selectedConfig)
+
+	// Display files in selected config
+	files := indexFile.Configs[selectedConfig].Files
+	var selectedFile string
+	fileOptions := make([]huh.Option[string], 0, len(files))
+	for _, file := range files {
+		// Remove any surrounding quotes and use only the base filename
+		cleanFile := strings.Trim(file, "\"")
+		baseFile := filepath.Base(cleanFile)
+		fileOptions = append(fileOptions, huh.NewOption(baseFile, cleanFile))
+	}
+
+	log.Info("Presenting file selection to user", "optionCount", len(fileOptions))
+	form = huh.NewForm(
+		huh.NewGroup(
+			huh.NewSelect[string]().
+				Title("Select a file").
+				Options(fileOptions...).
+				Value(&selectedFile),
+		),
+	)
+
+	err = form.Run()
+	if err != nil {
+		log.Error("Error selecting file", "error", err)
+		return
+	}
+	log.Info("User selected file", "selectedFile", selectedFile)
+
+	// Read and display file contents
+	content, err := os.ReadFile(filepath.Clean(selectedFile))
+	if err != nil {
+		log.Error("Error reading file", "error", err)
+		return
+	}
+
+	fmt.Printf("Contents of %s:\n\n%s\n\n", filepath.Base(selectedFile), string(content))
+
+	// Prompt user to run 00-init.sh
+	var confirmRun bool
+	form = huh.NewForm(
+		huh.NewGroup(
+			huh.NewConfirm().
+				Title("Do you want to run 00-init.sh to provision the cluster?").
+				Value(&confirmRun),
+		),
+	)
+
+	err = form.Run()
+	if err != nil {
+		log.Error("Error in confirmation prompt", "error", err)
+		return
+	}
+
+	if confirmRun {
+		// Summarize the provision cluster process
+		fmt.Println("\nProvisioning Cluster Summary:")
+		fmt.Printf("- Configuration: %s\n", selectedConfig)
+		fmt.Printf("- Init Script: %s\n", filepath.Join(filepath.Dir(selectedFile), "00-init.sh"))
+		fmt.Printf("- Kubefirst Script: %s\n", filepath.Join(filepath.Dir(selectedFile), "01-kubefirst-cloud.sh"))
+		fmt.Printf("- Environment File: %s\n", filepath.Join(filepath.Dir(selectedFile), ".local.cloud.env"))
+
+		// Final confirmation
+		var finalConfirm bool
+		form = huh.NewForm(
+			huh.NewGroup(
+				huh.NewConfirm().
+					Title("Do you want to proceed with provisioning the cluster?").
+					Value(&finalConfirm),
+			),
+		)
+
+		err = form.Run()
+		if err != nil {
+			log.Error("Error in final confirmation prompt", "error", err)
+			return
+		}
+
+		if finalConfirm {
+			log.Info("User confirmed cluster provisioning")
+			fmt.Println("Provisioning cluster...")
+			// Add logic here to run 00-init.sh
+			// This is where you would implement the actual cluster provisioning
+			// For example:
+			// err := runInitScript(filepath.Join(filepath.Dir(selectedFile), "00-init.sh"))
+			// if err != nil {
+			//     log.Error("Error provisioning cluster", "error", err)
+			//     return
+			// }
+			// fmt.Println("Cluster provisioned successfully!")
+		} else {
+			log.Info("User cancelled cluster provisioning")
+			fmt.Println("Cluster provisioning cancelled.")
+		}
+	} else {
+		log.Info("User chose not to run 00-init.sh")
+		fmt.Println("Cluster provisioning cancelled.")
+	}
+}
+
+// Add any additional cluster-related functions here
+// For example:
+// func deleteCluster() { ... }
+// func listClusters() { ... }
+// func updateCluster() { ... }
